oucit/requirements.txt
Django==1.11.9
django-autocomplete-light==3.2.10
django-querysetsequence==0.8
django-ranged-response==0.2.0
django-simple-captcha==0.5.6
Pillow==5.0.0
PyMySQL==0.8.0
pytz==2017.3
six==1.11.0
typing==3.6.4
